LayoutTests/http/tests/xmlhttprequest/access-control-basic-blacklist-request-headers-expected.txt

CONSOLE MESSAGE: line 1: Refused to set unsafe header "accept-charset"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "accept-encoding"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "connection"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "content-length"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "content-transfer-encoding"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "date"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "expect"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "host"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "keep-alive"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "referer"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "te"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "trailer"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "transfer-encoding"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "upgrade"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "via"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "Proxy-"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "Proxy-test"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "PROXY-FOO"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "Sec-"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "Sec-test"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "SEC-FOO"
CONSOLE MESSAGE: line 1: Refused to send header "origin" cross-domain.
CONSOLE MESSAGE: line 1: Refused to send header "authorization" cross-domain.
CONSOLE MESSAGE: line 1: Refused to send header "cookie" cross-domain.
CONSOLE MESSAGE: line 1: Refused to send header "cookie2" cross-domain.
Test that setRequestHeader cannot be used to send security-sensitive headers cross-site.

SUCCESS
